Move some types and traits from cache module to dedicated modules
