+++ src/lib.rs
@@ -124,8 +124,4 @@
 
 pub mod sync;
 
-pub(crate) mod common {
-    pub(crate) mod deque;
-    pub(crate) mod frequency_sketch;
-    pub(crate) mod thread_pool;
-}
+pub(crate) mod common;
